Import refactoring and fix the circular import err
