+++ src/api/services/__init__.py
@@ -1,7 +1,10 @@
 from .admin_service import AdminService
+from .analytics import AnalyticsService
 from .base import ContentService
 from .category import CategoryService
 from .external_site_user import ExternalSiteUserService
+from .health_check import HealthCheckService
+from .messages import TelegramNotificationService
 from .task import TaskService
 
 __all__ = (
@@ -10,4 +13,7 @@ __all__ = (
     "TaskService",
     "ExternalSiteUserService",
     "AdminService",
+    "AnalyticsService",
+    "HealthCheckService",
+    "TelegramNotificationService",
 )
