+++ src/bot/services/__init__.py
@@ -1,0 +1,3 @@
+from .unsubscribe_reason import UnsubscribeReasonService
+
+__all__ = ("UnsubscribeReasonService",)
