+++ src/core/db/repository/__init__.py
@@ -1,3 +1,4 @@
+from .admin_repository import AdminUserRepository
 from .base import AbstractRepository, ContentRepository
 from .category import CategoryRepository
 from .external_site_user import ExternalSiteUserRepository
@@ -13,4 +14,5 @@ __all__ = (
     "UserRepository",
     "ExternalSiteUserRepository",
     "UnsubscribeReasonRepository",
+    "AdminUserRepository",
 )
