+++ src/main.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI
 
 from src.api import init_app
-from src.bot import init_bot
-from src.bot.bot import shutdown_bot, startup_bot
+from src.bot import shutdown_bot, startup_bot
+from src.bot.handlers import init_bot
 from src.core.utils import set_ngrok
 from src.depends import Container
